Updated sources to match MSPM0 SDK 2.06.00.05 w/o TI Commercial Comp

+++ examples/nortos/LP_MSPM0C1104/bsl/bsl_host_mcu_to_mspm0c110x_target_uart/bsl_uart.c
@@ -237,7 +237,7 @@ BSL_error_t Host_BSL_writeMemory(
     uint16_t ui16PayloadSize;
     uint16_t ui16PacketSize;
     uint16_t ui16CRC;
-    uint16_t ui16BytesToWrite = len;
+    uint32_t ui16BytesToWrite = len;
     uint32_t TargetAddress    = addr;
 
     //  pSection->checksum = softwareCRC(pSection->pMemory, pSection->mem_size);
