+++ examples/nortos/LP_MSPM0G3507/bsl/bsl_host_mcu_to_mspm0g1x0x_g3x0x_target_spi/bsl_spi.c
@@ -242,7 +242,7 @@ BSL_error_t Host_BSL_writeMemory(
     uint16_t ui16PayloadSize;
     uint16_t ui16PacketSize;
     uint32_t ui32CRC;
-    uint16_t ui16BytesToWrite = len;
+    uint32_t ui16BytesToWrite = len;
     uint32_t TargetAddress    = addr;
 
     //  pSection->checksum = softwareCRC(pSection->pMemory, pSection->mem_size);
